backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.3
python-multipart==0.0.6
pydantic==2.5.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
huggingface-hub==0.16.4
pgvector==0.2.4
sqlalchemy==2.0.23 